requirements.txt
# Core dependencies
fastapi>=0.115.12      # Latest release Mar 23, 2025 :contentReference[oaicite:0]{index=0}
uvicorn>=0.34.2        # Latest release Apr 19, 2025 :contentReference[oaicite:1]{index=1}
httpx>=0.27.2          # Latest release Feb 14, 2025 :contentReference[oaicite:2]{index=2}
pydantic>=2.11.1       # Latest release Mar 28, 2025 :contentReference[oaicite:3]{index=3}
protobuf>=4.24.7       # Latest release May 10, 2025 :contentReference[oaicite:4]{index=4}
grpcio>=1.63.0         # Latest release Apr 30, 2025 :contentReference[oaicite:5]{index=5}
grpcio-tools>=1.63.0   # Latest release Apr 30, 2025 :contentReference[oaicite:6]{index=6}

# Database and vector search
valkey-glide>=1.3.5       # Valkey client for Python
faiss-cpu>=1.11.0      # Latest release Feb 20, 2025
numpy>=1.25.2          # Latest release Mar 10, 2025

# Embeddings and LLM
sentence-transformers>=4.1.0  # Latest release Apr 15, 2025 :contentReference[oaicite:7]{index=7}
ollama>=0.4.8                 # Latest release Feb 28, 2025

# Testing and development
pytest>=8.2.0            # Latest release Mar 5, 2025
pytest-benchmark>=4.1.0  # Latest release Jan 10, 2025
hypothesis>=6.97.0       # Latest release Apr 2, 2025
ruff>=0.0.300            # Latest release May 1, 2025
black>=24.1.0            # Latest release Apr 22, 2025
mypy>=1.7.1              # Latest release Mar 18, 2025
pyrefly>=0.16.0

# TUI
textual>=0.55.0          # Latest release Mar 25, 2025
rich>=14.0.0             # Latest release Apr 8, 2025
# Utilities
pyyaml>=6.0.2              # Latest release Feb 14, 2025
python-dotenv>=1.0.1     # Latest release Mar 2, 2025

# Documentation
mkdocs>=1.5.3            # Documentation generator
mkdocs-material>=9.4.6   # Material theme for MkDocs
pymdown-extensions>=10.3 # Markdown extensions for MkDocs
